# docs/swagger.yaml
definitions:
  models.AuditData:
    properties:
      id:
        type: integer
      modificationDate:
        type: string
      natureOfModification:
        type: string
      ruleId:
        type: integer
      userId:
        type: integer
    type: object
...
  models.NotificationPreference:
    properties:
      exam_notification:
        type: boolean
      homework_notification:
        type: boolean
      social_notification:
        type: boolean
    required:
    - exam_notification
    - homework_notification
    - social_notification
    type: object
...
  models.NotificationTokens:
    properties:
      notifications:
        items:
          $ref: '#/definitions/models.NotificationToken'
        type: array
    type: object
...
  models.Rule:
    properties:
      ApplicationCondition:
        type: string
      Description:
        type: string
      Title:
        type: string
      effectiveDate:
        type: string
      id:
        type: integer
    required:
    - ApplicationCondition
    - Description
    - Title
    type: object
  models.RuleModify:
    properties:
      ApplicationCondition:
        type: string
      Description:
        type: string
      Title:
        type: string
    type: object
...
  utils.HTTPError:
    properties:
      code:
        type: integer
      error:
        type: string
      title:
        type: string
    type: object
...
host: https://user-api-production-99c2.up.railway.app/
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: User API
  version: "1.0"
paths:
...
  /rules:
    get:
      consumes:
      - application/json
      description: Returns a list of all rules in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of rules
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Rule'
              type: array
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Get all rules
      tags:
      - Rules
    post:
      consumes:
      - application/json
      description: Create a new rule
      parameters:
      - description: Rule creation Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Rule'
      produces:
      - text/plain
      responses:
        "201":
          description: Rule created correctly
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Create a new rule
      tags:
      - Rules
  /rules/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a user from the database
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Rule successfully deleted
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Delete user by ID
      tags:
      - Rules
    put:
      consumes:
      - application/json
      description: Updates the contents of a rule
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Elements to modify
        in: body
        name: modifications
        required: true
        schema:
          $ref: '#/definitions/models.RuleModify'
      responses:
        "200":
          description: rule updated successfully
        "400":
          description: Invalid user ID format or request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Modify rule password
      tags:
      - Rules
  /rules/audit:
    get:
      consumes:
      - application/json
      description: Returns a list of all audits in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of audits
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.AuditData'
              type: array
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Get all audits
      tags:
      - Rules
  /users:
    get:
      consumes:
      - application/json
      description: Returns a list of all users in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.User'
              type: array
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - Users
  /users/:id/notifications/preference:
    get:
      consumes:
      - application/json
      description: Modify the preference of a notification type between exam_notification
        homework_notification or social_notification
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: preferences
          schema:
            $ref: '#/definitions/models.NotificationPreference'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Modify the preference of a notification type
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Modify the preference of a notification type between exam_notification
        homework_notification or social_notification
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: NotificationPreferenceRequest payload
        in: body
        name: NotificationPreferenceRequest
        required: true
        schema:
          $ref: '#/definitions/models.NotificationPreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: preference changed successfully
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Modify the preference of a notification type
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a user from the database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User successfully deleted
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Delete user by ID
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Returns a specific user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            additionalProperties:
              $ref: '#/definitions/models.User'
            type: object
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates user information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateDto'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user data
          schema:
            additionalProperties:
              $ref: '#/definitions/models.UserUpdateDto'
            type: object
        "400":
          description: Invalid user ID format or request format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Modify user
      tags:
      - Users
  /users/{id}/notifications:
    get:
      consumes:
      - application/json
      description: Send a notification to users sent in body
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users notified successfully
          schema:
            $ref: '#/definitions/models.NotificationTokens'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Send a notification to users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Set a notification token to users in order to be able to send push
        notifications
      parameters:
      - description: NotificationSetUpRequest payload
        in: body
        name: NotificationToken
        required: true
        schema:
          $ref: '#/definitions/models.NotificationSetUpRequest'
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Token Setup successful
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Set a notification token to users
      tags:
      - Users
  /users/{id}/teacher:
    put:
      consumes:
      - application/json
      description: makes the user role "teacher"
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: User blocked successfully
          schema:
            type: string
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Make a user a "teacher"
      tags:
      - Users
  /users/block/{id}:
    put:
      consumes:
      - application/json
      description: Blocks a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: User blocked successfully
          schema:
            type: string
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      security:
      - Bearer: []
      summary: Block user
      tags:
      - Users
  /users/notify:
    post:
      consumes:
      - application/json
      description: Send a notification to users sent in body
      parameters:
      - description: NotificationToken payload
        in: body
        name: NotificationToken
        required: true
        schema:
          $ref: '#/definitions/models.NotifyRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Users notified successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Send a notification to users
      tags:
      - Users
...
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"